+++ packages/core/src/lib/types.ts
@@ -2,7 +2,6 @@ import { Signal, WritableSignal } from '@angular/core';
 
 import { AsyncQueryMarker, AsyncQuerySignal } from './markers/async-query';
 import { AsyncSourceMarker, AsyncSourceSignal } from './markers/async-source';
-import { AsyncStreamMarker, AsyncStreamSignal } from './markers/async-stream';
 import type { EntityLoaderSurface } from './markers/entity-loader';
 import { FormMarker, FormSignal } from './markers/form';
 import { StatusMarker, StatusSignal } from './markers/status';
@@ -213,8 +212,6 @@ export type TreeNode<T> = {
     ? AsyncSourceSignal<V>
     : T[K] extends AsyncQueryMarker<infer In, infer Out>
     ? AsyncQuerySignal<In, Out>
-    : T[K] extends AsyncStreamMarker<infer Chunk, infer State>
-    ? AsyncStreamSignal<Chunk, State>
     : T[K] extends Primitive
     ? CallableWritableSignal<T[K]>
     : T[K] extends readonly unknown[]
@@ -1015,8 +1012,6 @@ export type DeepEntityAwareTreeNode<T> = {
     ? AsyncSourceSignal<V>
     : T[K] extends AsyncQueryMarker<infer In, infer Out>
     ? AsyncQuerySignal<In, Out>
-    : T[K] extends AsyncStreamMarker<infer Chunk, infer State>
-    ? AsyncStreamSignal<Chunk, State>
     : T[K] extends object
     ? DeepEntityAwareTreeNode<T[K]>
     : CallableWritableSignal<T[K]>;
@@ -1044,8 +1039,6 @@ export type EntityAwareTreeNode<T> = {
     ? AsyncSourceSignal<V>
     : T[K] extends AsyncQueryMarker<infer In, infer Out>
     ? AsyncQuerySignal<In, Out>
-    : T[K] extends AsyncStreamMarker<infer Chunk, infer State>
-    ? AsyncStreamSignal<Chunk, State>
     : CallableWritableSignal<T[K]>;
 };